+++ src/platform/Linux/CHIPLinuxStorage.cpp
@@ -230,7 +230,7 @@ CHIP_ERROR ChipLinuxStorage::WriteValueStr(const char * key, const char * val)
 
 CHIP_ERROR ChipLinuxStorage::WriteValueBin(const char * key, const uint8_t * data, size_t dataLen)
 {
-    static const size_t kMaxBlobSize = 10 * 1024;
+    static const size_t kMaxBlobSize = static_cast<size_t>(10) * 1024;
 
     CHIP_ERROR retval = CHIP_NO_ERROR;
     chip::Platform::ScopedMemoryBuffer<char> encodedData;
